Admin and Editor roles

+++ routes/web.php
@@ -17,9 +17,9 @@ use Illuminate\Support\Facades\Route;
 Auth::routes();
 
 Route::get('', [NewsController::class, 'index'])->name('news');
-Route::get('news/create', [NewsController::class, 'create'])->name('news.create');
-Route::post('news/create', [NewsController::class, 'store'])->name('news.store');
+Route::get('news/create', [NewsController::class, 'create'])->name('news.create')->middleware('role:admin');
+Route::post('news/create', [NewsController::class, 'store'])->name('news.store')->middleware('role:admin');
 Route::get('news/{news}', [NewsController::class, 'show'])->name('news.show');
 Route::get('news/edit/{news}', [NewsController::class, 'edit'])->name('news.edit');
 Route::post('news/edit/{news}', [NewsController::class, 'update'])->name('news.update');
-Route::delete('news/delete/{news}', [NewsController::class, 'destroy'])->name('news.delete');
+Route::delete('news/delete/{news}', [NewsController::class, 'destroy'])->name('news.delete')->middleware('role:admin');
